taken out state management and move to app.jsx
