+++ src/Feed.jsx
@@ -2,23 +2,13 @@ import { useState } from "react";
 import { Post } from "./Post";
 import { useEffect } from "react";
 
-export const Feed = () => {
-  // here we are using the useState hook to be able to change the data
-  const [posts, setPosts] = useState([]);
+export const Feed = (props) => {
 
-  // the useEffect hook allows us to run code once when the component is rendered
-  useEffect(() => {
-    // here we are fetching the data from the API, code snippet I got from the weather app project and adjusted
-    fetch("https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts")
-      .then((response) => response.json())
-      .then((posts) => {
-        setPosts(posts);
-      });
-  }, []);
+  // props is an object that contains all the properties that were passed to the component
 
   return (
     <div className="feed">
-      {posts.map((post) => (
+      {props.posts.map((post) => (
         // here we are rendering the Post component
         <Post
           key={post._id}
