+++ src/Form.jsx
@@ -1,28 +1,11 @@
 import { useState } from "react";
 
 // Here is where we create the form/input field
-export const Form = () => {
-  const [message, setMessage] = useState("");
+export const Form = (props) => {
 
-  const onChange = (event) => {
-    setMessage(event.target.value);
-  };
-
-  const onSubmit = (event) => {
-    event.preventDefault();
-    // URL from https://github.com/Technigo/project-happy-thoughts-vite/blob/main/instructions.md#create-a-thought
-    const url = "https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts";
-    fetch(url, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ message }),
-    });
-  };
 
   return (
-    <form className="form" onSubmit={onSubmit}>
+    <form className="form" onSubmit={props.onSubmit}>
       <label className="label" htmlFor="message">
         What's making you happy right now?
       </label>
@@ -30,7 +13,7 @@ export const Form = () => {
         className="textInput"
         id="message"
         rows="3"
-        onChange={onChange}
+        onChange={props.onChange}
       ></textarea>
       <button className="submitButton" type="submit">
         ❤️ Send Happy Thought ❤️
